requirements.txt
pandas>=1.3.0
requests>=2.26.0
beautifulsoup4>=4.9.3
selenium>=4.0.0
python-dotenv>=0.19.0
webdriver-manager>=4.0.0
psycopg2-binary>=2.9.3
